Fixed the run code button
And refactored debug code button to align with run code (functionally the same)
