+++ js_client/webpack.config.js
@@ -11,6 +11,7 @@ module.exports = {
         robot_3d_plot: './src/robot3dplot.png',
         dark_mode: './src/Toolbox/auto_dark_mode.ts',
         inspectionPoints: './src/interaction/inspectionPoints.ts',
+        runCode: './src/interaction/runCode.ts',
         monacoEditor: './src/monacoExperiment.ts'
     },
     devtool: 'inline-source-map',
